+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 # Unreleased
 * Add `AchReturnCode` to transaction object
+* Add ExchangeRateQuote API
 
 ## 3.14.0 
 * Add `PaymentMethodCustomerDataUpdated` webhook notification support
-* Add ExchangeRateQuote API
 
 ## 3.13.0
 * Make `junit-vintage` a test dependency (thanks @jamesbroadhead)